Move component TestedNumberCard from 'components/cards' to 'components/index/CardsReference'
